backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
qrcode[pil]==7.4.2
django-filter==23.5
python-decouple==3.8
Pillow==10.2.0
django-extensions==3.2.3
gunicorn==21.2.0
dj-database-url==2.1.0
drf-spectacular==0.27.1
django-debug-toolbar==4.3.0

# Testing dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-factoryboy==2.6.0
factory-boy==3.3.0
faker==20.1.0
django-environ==0.10.0
qrcode[pil]
